+++ x-pack/plugins/triggers_actions_ui/public/application/components/builtin_action_types/swimlane/swimlane.tsx
@@ -7,6 +7,7 @@
 
 import { isEmpty } from 'lodash';
 import { lazy } from 'react';
+import { i18n } from '@kbn/i18n';
 import {
   ActionTypeModel,
   ConnectorValidationResult,
@@ -18,9 +19,22 @@ import {
   SwimlaneSecrets,
   SwimlaneActionParams,
 } from './types';
-import * as i18n from './translations';
 import { isValidUrl } from '../../../lib/value_validators';
 import { validateMappingForConnector } from './helpers';
+
+export const SW_SELECT_MESSAGE_TEXT = i18n.translate(
+  'xpack.triggersActionsUI.components.builtinActionTypes.swimlaneAction.selectMessageText',
+  {
+    defaultMessage: 'Create record in Swimlane',
+  }
+);
+
+export const SW_ACTION_TYPE_TITLE = i18n.translate(
+  'xpack.triggersActionsUI.components.builtinActionTypes.swimlaneAction.actionTypeTitle',
+  {
+    defaultMessage: 'Create Swimlane Record',
+  }
+);
 
 export function getActionType(): ActionTypeModel<
   SwimlaneConfig,
@@ -30,11 +44,12 @@ export function getActionType(): ActionTypeModel<
   return {
     id: '.swimlane',
     iconClass: lazy(() => import('./logo')),
-    selectMessage: i18n.SW_SELECT_MESSAGE_TEXT,
-    actionTypeTitle: i18n.SW_ACTION_TYPE_TITLE,
+    selectMessage: SW_SELECT_MESSAGE_TEXT,
+    actionTypeTitle: SW_ACTION_TYPE_TITLE,
     validateConnector: async (
       action: SwimlaneActionConnector
     ): Promise<ConnectorValidationResult<SwimlaneConfig, SwimlaneSecrets>> => {
+      const translations = await import('./translations');
       const configErrors = {
         apiUrl: new Array<string>(),
         appId: new Array<string>(),
@@ -51,19 +66,22 @@ export function getActionType(): ActionTypeModel<
       };
 
       if (!action.config.apiUrl) {
-        configErrors.apiUrl = [...configErrors.apiUrl, i18n.SW_API_URL_REQUIRED];
+        configErrors.apiUrl = [...configErrors.apiUrl, translations.SW_API_URL_REQUIRED];
       } else if (action.config.apiUrl) {
         if (!isValidUrl(action.config.apiUrl)) {
-          configErrors.apiUrl = [...configErrors.apiUrl, i18n.SW_API_URL_INVALID];
+          configErrors.apiUrl = [...configErrors.apiUrl, translations.SW_API_URL_INVALID];
         }
       }
 
       if (!action.secrets.apiToken) {
-        secretsErrors.apiToken = [...secretsErrors.apiToken, i18n.SW_REQUIRED_API_TOKEN_TEXT];
+        secretsErrors.apiToken = [
+          ...secretsErrors.apiToken,
+          translations.SW_REQUIRED_API_TOKEN_TEXT,
+        ];
       }
 
       if (!action.config.appId) {
-        configErrors.appId = [...configErrors.appId, i18n.SW_REQUIRED_APP_ID_TEXT];
+        configErrors.appId = [...configErrors.appId, translations.SW_REQUIRED_APP_ID_TEXT];
       }
 
       const mappingErrors = validateMappingForConnector(
@@ -80,6 +98,7 @@ export function getActionType(): ActionTypeModel<
     validateParams: async (
       actionParams: SwimlaneActionParams
     ): Promise<GenericValidationResult<unknown>> => {
+      const translations = await import('./translations');
       const errors = {
         'subActionParams.incident.ruleName': new Array<string>(),
         'subActionParams.incident.alertId': new Array<string>(),
@@ -91,11 +110,11 @@ export function getActionType(): ActionTypeModel<
       const hasIncident = actionParams.subActionParams && actionParams.subActionParams.incident;
 
       if (hasIncident && !actionParams.subActionParams.incident.ruleName?.length) {
-        errors['subActionParams.incident.ruleName'].push(i18n.SW_REQUIRED_RULE_NAME);
+        errors['subActionParams.incident.ruleName'].push(translations.SW_REQUIRED_RULE_NAME);
       }
 
       if (hasIncident && !actionParams.subActionParams.incident.alertId?.length) {
-        errors['subActionParams.incident.alertId'].push(i18n.SW_REQUIRED_ALERT_ID);
+        errors['subActionParams.incident.alertId'].push(translations.SW_REQUIRED_ALERT_ID);
       }
 
       return validationResult;
